Add ability to use Messenger in Conversations
Previously Messenger would only permit sending to CommandSenders despite
the Bukkit API also containing Conversations. If Messenger was used in a
Conversation, nothing would appear to happen due to the filtering done by
Bukkit.

With this commit it is now possible to send Conversables messages via the
sendRaw() methods. This will default to using sendMessage() if the
provided CommandSender is not a Conversable type. If it is however, the
new methods will use sendRawMessage() instead.